aplique la base de datos
