--- src/environments/environment.ts
@@ -1,12 +1,0 @@
-export const environment = {
-  firebase: {
-    projectId: 'my-trainer-842fa',
-    appId: '1:27757548967:web:685ad6943ad52d63351282',
-    storageBucket: 'my-trainer-842fa.appspot.com',
-    locationId: 'southamerica-east1',
-    apiKey: 'AIzaSyAqwcWVbzAN_jxdAm_5u3uPg3uFbEUu4A4',
-    authDomain: 'my-trainer-842fa.firebaseapp.com',
-    messagingSenderId: '27757548967',
-    measurementId: 'G-37WXQL8TL6',
-  },
-};
